+++ src/L10NSharp/XLiffUtils/XLiffTransUnitUpdater.cs
@@ -19,13 +19,13 @@ namespace L10NSharp.XLiffUtils
 		// with the value of kOSNewline.
 		internal string _literalNewline = "\\n";
 
-		private readonly XLiffLocalizedStringCache _stringCache;
+		private readonly XliffLocalizedStringCache _stringCache;
 		private readonly string                    _defaultLang;
 		private          bool                      _updated;
 
 
 		/// ------------------------------------------------------------------------------------
-		internal XLiffTransUnitUpdater(XLiffLocalizedStringCache cache)
+		internal XLiffTransUnitUpdater(XliffLocalizedStringCache cache)
 		{
 			_stringCache = cache;
 			_defaultLang = LocalizationManager.kDefaultLang;