+++ src/L10NSharpTests/LocalizationManagerTestsBase.cs
@@ -245,7 +245,7 @@ namespace L10NSharp.Tests
 		{
 			using(var folder = new TempFolder())
 			{
-				SetupManager(folder, "en");
+				SetupManager(folder);
 				//This was the original assertion, and it worked:
 				//   Assert.AreEqual("from English Translation", LocalizationManager.GetDynamicString(AppId, "theId", "some default"));
 				//However, later I decided, I don't care what is in the English Translation, either. If the c# code just gave a new
@@ -267,7 +267,7 @@ namespace L10NSharp.Tests
 		{
 			using(var folder = new TempFolder())
 			{
-				SetupManager(folder, "en");
+				SetupManager(folder);
 				Assert.AreEqual("from c# code", LocalizationManager.GetDynamicString(AppId, "theId", "from c# code"));
 			}
 		}
@@ -287,7 +287,7 @@ namespace L10NSharp.Tests
 		{
 			using (var folder = new TempFolder())
 			{
-				SetupManager(folder, "en");
+				SetupManager(folder);
 				Assert.That(LocalizationManager.GetDynamicString(AppId, "blahId", null), Is.EqualTo("blah"), "With no default supplied, should find saved English");
 			}
 		}
@@ -297,7 +297,7 @@ namespace L10NSharp.Tests
 		{
 			using (var folder = new TempFolder())
 			{
-				SetupManager(folder, "en");
+				SetupManager(folder);
 				Assert.That(LocalizationManager.GetDynamicString(AppId, "blahId", null), Is.EqualTo("blah"), "With no default supplied, should find saved English");
 				using (var extra = CreateLocalizationManager("nonsense", "more nonsense", "1.0"))
 				{
@@ -326,17 +326,22 @@ namespace L10NSharp.Tests
 			}
 		}
 
-		//NOTE: the TestName parameter is only here to work around an NUnit bug in which
-		//NUnit doesn't run all the test cases when some differ only by the values in an array parameter
-		//cases where we expect to get back the english in the code
-		[TestCase(new[] { "en" }, "blahInEnglishCode", "en", TestName = "GetString_OverloadThatTakesListOfLanguages_Works_1")]
-		[TestCase(new[] { "en", "fr" }, "blahInEnglishCode", "en", TestName = "GetString_OverloadThatTakesListOfLanguages_Works_2")]
-		[TestCase(new[] { "ar", "en" }, "blahInEnglishCode", "en", TestName = "GetString_OverloadThatTakesListOfLanguages_Works_3")] // our arabic doesn't have a translation of 'blah', so fall to the code's English
-		[TestCase(new[] { "zz", "en", "fr" }, "blahInEnglishCode", "en", TestName = "GetString_OverloadThatTakesListOfLanguages_Works_4")]
-		//cases where we expect to get back the French
-		[TestCase(new[] { "fr" }, "blahInFrench", "fr", TestName = "GetString_OverloadThatTakesListOfLanguages_Works_5")]
-		[TestCase(new[] { "fr", "en" }, "blahInFrench", "fr", TestName = "GetString_OverloadThatTakesListOfLanguages_Works_6")]
-		[TestCase(new[] { "ar", "fr", "en" }, "blahInFrench", "fr", TestName = "GetString_OverloadThatTakesListOfLanguages_Works_7")] // our arabic doesn't have a translation of 'blah', so fall to French
+		/// <summary>
+		/// Tests that GetString returns the first language in order of preference with an available string
+		/// </summary>
+		/// <remarks>
+		/// NOTE: the TestName parameter is only here to work around an NUnit bug in which
+		/// NUnit doesn't run all the test cases when some differ only by the values in an array parameter.
+		/// </remarks>
+		//cases where we expect to get back the english in the code:
+		[TestCase(new[] { "en" }, "blahInEnglishCode", "en", TestName = "en; finds en")]
+		[TestCase(new[] { "en", "fr" }, "blahInEnglishCode", "en", TestName = "en,fr; finds en")]
+		[TestCase(new[] { "ar", "en" }, "blahInEnglishCode", "en", TestName = "ar,en; finds en")] // our arabic doesn't have a translation of 'blah', so fall back to the code's English
+		[TestCase(new[] { "zz", "en", "fr" }, "blahInEnglishCode", "en", TestName = "zz,en,fr; finds en")]
+		//cases where we expect to get back the French:
+		[TestCase(new[] { "fr" }, "blahInFrench", "fr", TestName = "fr; finds fr")]
+		[TestCase(new[] { "fr", "en" }, "blahInFrench", "fr", TestName = "fr,en; finds fr")]
+		[TestCase(new[] { "ar", "fr", "en" }, "blahInFrench", "fr", TestName = "ar,fr,en; finds fr")] // our arabic doesn't have a translation of 'blah', so fall back to French
 		public void GetString_OverloadThatTakesListOfLanguages_Works(IEnumerable<string> preferredLangIds,  string expectedResult, string expectedLanguage)
 		{
 			using(var folder = new TempFolder())
@@ -397,11 +402,6 @@ namespace L10NSharp.Tests
 			}
 		}
 
-		int compareCultureTags(CultureInfo first, CultureInfo second)
-		{
-			return first.IetfLanguageTag.CompareTo(second.IetfLanguageTag);
-		}
-
 		[Test]
 		public void GetDynamicStringInEnglish_NoDefault_FindsEnglishWithFolders()
 		{
@@ -420,17 +420,22 @@ namespace L10NSharp.Tests
 			}
 		}
 
-		//NOTE: the TestName parameter is only here to work around an NUnit bug in which
-		//NUnit doesn't run all the test cases when some differ only by the values in an array parameter
-		//cases where we expect to get back the english in the code
-		[TestCase(new[] { "en" }, "blahInEnglishCode", "en", TestName = "GetString_OverloadThatTakesListOfLanguages_WorksWithFolders_1")]
-		[TestCase(new[] { "en", "fr" }, "blahInEnglishCode", "en", TestName = "GetString_OverloadThatTakesListOfLanguages_WorksWithFolders_2")]
-		[TestCase(new[] { "ar", "en" }, "blahInEnglishCode", "en", TestName = "GetString_OverloadThatTakesListOfLanguages_WorksWithFolders_3")] // our arabic doesn't have a translation of 'blah', so fall to the code's English
-		[TestCase(new[] { "zz", "en", "fr" }, "blahInEnglishCode", "en", TestName = "GetString_OverloadThatTakesListOfLanguages_WorksWithFolders_4")]
-		//cases where we expect to get back the French
-		[TestCase(new[] { "fr" }, "blahInFrench", "fr", TestName = "GetString_OverloadThatTakesListOfLanguages_WorksWithFolders_5")]
-		[TestCase(new[] { "fr", "en" }, "blahInFrench", "fr", TestName = "GetString_OverloadThatTakesListOfLanguages_WorksWithFolders_6")]
-		[TestCase(new[] { "ar", "fr", "en" }, "blahInFrench", "fr", TestName = "GetString_OverloadThatTakesListOfLanguages_WorksWithFolders_7")] // our arabic doesn't have a translation of 'blah', so fall to French
+		/// <summary>
+		/// Tests that GetString returns the first language in order of preference with an available string
+		/// </summary>
+		/// <remarks>
+		/// NOTE: the TestName parameter is only here to work around an NUnit bug in which
+		/// NUnit doesn't run all the test cases when some differ only by the values in an array parameter.
+		/// </remarks>
+		//cases where we expect to get back the english in the code:
+		[TestCase(new[] { "en" }, "blahInEnglishCode", "en", TestName = "en; finds en in folders")]
+		[TestCase(new[] { "en", "fr" }, "blahInEnglishCode", "en", TestName = "en,fr; finds en in folders")]
+		[TestCase(new[] { "ar", "en" }, "blahInEnglishCode", "en", TestName = "ar,en; finds en in folders")] // our arabic doesn't have a translation of 'blah', so fall back to the code's English
+		[TestCase(new[] { "zz", "en", "fr" }, "blahInEnglishCode", "en", TestName = "zz,en,fr; finds en in folders")]
+		//cases where we expect to get back the French:
+		[TestCase(new[] { "fr" }, "blahInFrench", "fr", TestName = "fr; finds fr in folders")]
+		[TestCase(new[] { "fr", "en" }, "blahInFrench", "fr", TestName = "fr,en; finds fr in folders")]
+		[TestCase(new[] { "ar", "fr", "en" }, "blahInFrench", "fr", TestName = "ar,fr,en; finds fr in folders")] // our arabic doesn't have a translation of 'blah', so fall back to French
 		public void GetString_OverloadThatTakesListOfLanguages_WorksWithFolders(IEnumerable<string> preferredLangIds,  string expectedResult, string expectedLanguage)
 		{
 			LocalizationManager.UseLanguageCodeFolders = true;
@@ -461,7 +466,7 @@ namespace L10NSharp.Tests
 		/// <summary>
 		/// - "Installs" English, Arabic, and French
 		/// - Sets the UI language
-		/// - Constructs a LocalizationManager and adds it to LocalizationManagerInternal<T>.LoadedManagers[AppId]
+		/// - Constructs a LocalizationManager and adds it to LocalizationManagerInternal&lt;T&gt;.LoadedManagers[AppId]
 		/// </summary>
 		public void SetupManager(TempFolder folder, string uiLanguageId = LocalizationManager.kDefaultLang)
 		{
@@ -566,6 +571,52 @@ namespace L10NSharp.Tests
 				CreateTransUnitVariant("es", "bleah"));
 			spanishDoc.AddTransUnit(tu3);
 			spanishDoc.Save(Path.Combine(folderPath, LocalizationManager.GetTranslationFileNameForLanguage(AppId, "es")));
+		}
+
+		private void AddChineseOfChinaTranslation(string folderPath)
+		{
+			var chineseDoc = CreateNewDocument(null, "en", "zh-CN");
+			// first unit
+			var tu = CreateTransUnit("theId", false,
+				CreateTransUnitVariant("en", "from English Translation"),
+				CreateTransUnitVariant("zh-CN", "from Chinese (China) Translation"),
+				"Test", TranslationStatus.Approved);
+			chineseDoc.AddTransUnit(tu);
+			// second unit
+			var tu2 = CreateTransUnit("notUsedId", false,
+				CreateTransUnitVariant("en", "no longer used English text"),
+				CreateTransUnitVariant("zh-CN", "no longer used Chinese (China) text"),
+				null, TranslationStatus.Approved);
+			chineseDoc.AddTransUnit(tu2);
+			// third unit
+			var tu3 = CreateTransUnit("blahId", false,
+				CreateTransUnitVariant("en", "blah"),
+				CreateTransUnitVariant("zh-CN", "中文(中国) blah"));
+			chineseDoc.AddTransUnit(tu3);
+			chineseDoc.Save(Path.Combine(folderPath, LocalizationManager.GetTranslationFileNameForLanguage(AppId, "zh-CN")));
+		}
+
+		private void AddChineseOfTaiwanTranslation(string folderPath)
+		{
+			var chineseDoc = CreateNewDocument(null, "en", "zh-TW");
+			// first unit
+			var tu = CreateTransUnit("theId", false,
+				CreateTransUnitVariant("en", "from English Translation"),
+				CreateTransUnitVariant("zh-TW", "from Chinese (Taiwan) Translation"),
+				"Test", TranslationStatus.Approved);
+			chineseDoc.AddTransUnit(tu);
+			// second unit
+			var tu2 = CreateTransUnit("notUsedId", false,
+				CreateTransUnitVariant("en", "no longer used English text"),
+				CreateTransUnitVariant("zh-TW", "no longer used Chinese (Taiwan) text"),
+				null, TranslationStatus.Approved);
+			chineseDoc.AddTransUnit(tu2);
+			// third unit
+			var tu3 = CreateTransUnit("blahId", false,
+				CreateTransUnitVariant("en", "blah"),
+				CreateTransUnitVariant("zh-TW", "中文(Taiwan) blah"));
+			chineseDoc.AddTransUnit(tu3);
+			chineseDoc.Save(Path.Combine(folderPath, LocalizationManager.GetTranslationFileNameForLanguage(AppId, "zh-TW")));
 		}
 
 		protected void AddRandomTranslation(string langId, string folderPath)
@@ -681,7 +732,7 @@ namespace L10NSharp.Tests
 		{
 			using (var folder = new TempFolder())
 			{
-				SetupManager(folder, "en");
+				SetupManager(folder);
 				var lm = LocalizationManagerInternal<T>.LoadedManagers.Values.First();
 				var tags = lm.GetAvailableUILanguageTags().ToArray();
 				Assert.That(tags.Length, Is.EqualTo(4));
@@ -895,38 +946,106 @@ namespace L10NSharp.Tests
 
 				// Check that we return the provided string for English.
 				var str = LocalizationManager.GetString("theId", "This is a test!", "This is only a test?", new[]{ "en", "fr", "ar" }, out var languageIdUsed);
-				Assert.AreEqual("This is a test!", str);
-				Assert.AreEqual("en", languageIdUsed);
+				Assert.That(str, Is.EqualTo("This is a test!"), "seeking en");
+				Assert.That(languageIdUsed, Is.EqualTo("en"), "seeking en");
 
 				// Check that asking for a specific form of English still returns the provided string.
 				str = LocalizationManager.GetString("theId", "This is a test!", "This is only a test?", new []{ "en-US", "es-ES", "fr-FR" }, out languageIdUsed);
-				Assert.AreEqual("This is a test!", str);
-				Assert.AreEqual("en", languageIdUsed);
+				Assert.That(str, Is.EqualTo("This is a test!"), "seeking en-US");
+				Assert.That(languageIdUsed, Is.EqualTo("en"), "seeking en-US");
 
 				// Check that we return the string from the second language when the first language doesn't have the string.
 				str = LocalizationManager.GetString("theId", "This is a test!", "This is only a test?", new[]{ "fr", "ar", "es" }, out languageIdUsed);
-				Assert.AreEqual("inArabic", str);
-				Assert.AreEqual("ar", languageIdUsed);
+				Assert.That(str, Is.EqualTo("inArabic"), "seeking fr, ar, es");
+				Assert.That(languageIdUsed, Is.EqualTo("ar"), "seeking fr, ar, es");
 
 				// Check that we return the string from the first language when it exists.
 				str = LocalizationManager.GetString("theId", "This is a test!", "This is only a test?", new []{ "es-ES", "en", "fr" }, out languageIdUsed);
-				Assert.AreEqual("from Spanish Translation", str);
-				Assert.AreEqual("es-ES", languageIdUsed);
+				Assert.That(str, Is.EqualTo("from Spanish Translation"), "seeking es-ES");
+				Assert.That(languageIdUsed, Is.EqualTo("es-ES"), "seeking es-ES");
 
 				// Check asking for the general form of the language when we have only a specific form.
 				str = LocalizationManager.GetString("theId", "This is a test!", "This is only a test?", new []{ "es", "en", "fr" }, out languageIdUsed);
-				Assert.AreEqual("from Spanish Translation", str);
-				Assert.AreEqual("es-ES", languageIdUsed);
+				Assert.That(str, Is.EqualTo("from Spanish Translation"), "seeking es");
+				Assert.That(languageIdUsed, Is.EqualTo("es-ES"), "seeking es");
 
 				// Check asking for a specific form of the language when we have only the general form.
 				str = LocalizationManager.GetString("theId", "This is a test!", "This is only a test?", new []{ "ar-AR", "en", "fr" }, out languageIdUsed);
-				Assert.AreEqual("inArabic", str);
-				Assert.AreEqual("ar", languageIdUsed);
+				Assert.That(str, Is.EqualTo("inArabic"), "seeking ar-AR");
+				Assert.That(languageIdUsed, Is.EqualTo("ar"), "seeking ar-AR");
 
 				// Check asking for a specific form of the language when we have only a different specific form.
 				str = LocalizationManager.GetString("theId", "This is a test!", "This is only a test?", new []{ "es-MX", "en-GB", "fr-FR" }, out languageIdUsed);
-				Assert.AreEqual("from Spanish Translation", str);
-				Assert.AreEqual("es-ES", languageIdUsed);
+				Assert.That(str, Is.EqualTo("from Spanish Translation"), "seeking es-MX");
+				Assert.That(languageIdUsed, Is.EqualTo("es-ES"), "seeking es-MX");
+			}
+		}
+
+		[Test]
+		public void TestMappingLanguageCodesToAvailable_FindsSpecificGivenGeneric()
+		{
+			LocalizationManager.SetUILanguage("en", true);
+			LocalizationManagerInternal<T>.LoadedManagers.Clear();
+			using (var folder = new TempFolder())
+			{
+				var installedFolder = Path.Combine(folder.Path, "installed");
+				AddEnglishTranslation(installedFolder, null);
+				AddChineseOfChinaTranslation(installedFolder);
+				LocalizationManagerInternal<T>.ChooseFallbackLanguage = (langTag, icon) =>
+					throw new NotImplementedException($"{langTag} shouldn't have stumped us");
+				var manager = LocalizationManager.Create("zh", AppId, AppName, AppVersion, installedFolder,
+					$"Temp/{Path.GetFileName(folder.Path)}/user", null, null, new string[] { });
+				LocalizationManagerInternal<T>.LoadedManagers[AppId] = (ILocalizationManagerInternal<T>)manager;
+
+				var langs = LocalizationManager.GetAvailableLocalizedLanguages();
+				Assert.That(langs, Is.EquivalentTo(new[] { "en", "zh-CN" }));
+
+				Assert.That(LocalizationManager.GetIsStringAvailableForLangId("theId", "zh"), Is.True, "zh should find zh-CN");
+				Assert.That(LocalizationManager.GetIsStringAvailableForLangId("theId", "zh-CN"), Is.True, "zh-CN should find zh-CN");
+				Assert.That(LocalizationManager.GetIsStringAvailableForLangId("theId", "zh-TW"), Is.True, "zh-TW should find zh-CN");
+				Assert.That(LocalizationManager.GetIsStringAvailableForLangId("theId", "en"), Is.True, "en should find en");
+
+				// Check asking for a specific form of the language when we have only a different specific form.
+				var str = LocalizationManager.GetString("theId", ".", "", new []{ "zh" }, out var languageIdUsed);
+				Assert.That(str, Is.EqualTo("from Chinese (China) Translation"));
+				Assert.That(languageIdUsed, Is.EqualTo("zh-CN"));
+			}
+		}
+
+		[Test]
+		public void TestMappingLanguageCodesToAvailable_AmbiguousOptions_PromptsUser([Values("zh-CN", "zh-TW")] string choice)
+		{
+			LocalizationManager.SetUILanguage("en", true);
+			LocalizationManagerInternal<T>.LoadedManagers.Clear();
+			using (var folder = new TempFolder())
+			{
+				var installedFolder = Path.Combine(folder.Path, "installed");
+				// ReSharper disable once AssignNullToNotNullAttribute
+				var userRelativeFolder = Path.Combine("Temp", Path.GetFileName(Path.GetDirectoryName(folder.Path)),
+					Path.GetFileName(folder.Path), "user");
+				AddEnglishTranslation(installedFolder, null);
+				AddChineseOfChinaTranslation(installedFolder);
+				AddChineseOfTaiwanTranslation(installedFolder);
+				var userPromptCount = 0;
+				LocalizationManagerInternal<T>.ChooseFallbackLanguage = (langTag, icon) =>
+				{
+					userPromptCount++;
+					Assert.That(langTag, Is.EqualTo("zh"));
+					return choice;
+				};
+				var manager = LocalizationManager.Create("zh", AppId, AppName, AppVersion, installedFolder,
+					userRelativeFolder, null, null, new string[] { });
+				Assert.That(userPromptCount, Is.EqualTo(1));
+				LocalizationManagerInternal<T>.LoadedManagers[AppId] = (ILocalizationManagerInternal<T>)manager;
+
+				var langs = LocalizationManager.GetAvailableLocalizedLanguages();
+				Assert.That(langs, Is.EquivalentTo(new[] { "en", "zh-CN", "zh-TW" }));
+				Assert.That(LocalizationManager.UILanguageId, Is.EqualTo(choice));
+
+				Assert.That(LocalizationManager.GetIsStringAvailableForLangId("theId", "zh"), Is.False, "zh is ambiguous");
+				Assert.That(LocalizationManager.GetIsStringAvailableForLangId("theId", "zh-CN"), Is.True, "zh-CN should find zh-CN");
+				Assert.That(LocalizationManager.GetIsStringAvailableForLangId("theId", "zh-TW"), Is.True, "zh-TW should find zh-TW");
+				Assert.That(LocalizationManager.GetIsStringAvailableForLangId("theId", "en"), Is.True, "en should find en");
 			}
 		}
 	}